+++ printf/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: mmorue <mmorue@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/31 15:35:55 by mmorue            #+#    #+#             */
-/*   Updated: 2023/02/27 17:41:10 by mmorue           ###   ########.fr       */
+/*   Updated: 2023/03/07 15:49:37 by mmorue           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 # include <unistd.h>
 # include <stdlib.h>
 
-int				ft_atoi(const char *str);
+long long		ft_atoi(const char *str);
 char			*ft_strdup(char *src);
 size_t			ft_strlen(const char *str);
 char			*ft_itoa(int n);
